+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## unreleased
 
+- Handle Filebeat overriding the "host" field with its own object by
+  normalizing the sensor name before rendering. If Filebeat is used,
+  the Suricata provided sensor name is lost, so use the Filebeat
+  provided host.name
+  instead. https://github.com/jasonish/evebox/issues/100
+
 [Full Changelog](https://github.com/jasonish/evebox/compare/0.10.2..master)
 
 ## 0.10.2 - 2019-01-30